Extract checkpointing
