+++ doc/source/api.rst
@@ -93,6 +93,7 @@ Functional Blox
    embedding.sale.update_sale
    embedding.sale.sample_actions_sale
    embedding.sale.sample_target_actions_sale
+   checkpointing.maybe_train_and_checkpoint
    target_net.soft_target_net_update
    target_net.hard_target_net_update
    cross_entropy_method.cem_sample
@@ -118,6 +119,7 @@ Data Blox
    embedding.sale.SALE
    embedding.sale.ActorSALE
    embedding.sale.CriticSALE
+   checkpointing.CheckpointState
    replay_buffer.ReplayBuffer
    replay_buffer.LAP
    function_approximator.mlp.MLP
